+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 * New rules: `no_debug_rule`, `no_empty_style_rule`, `no_loud_comment_rule`,
   `quote_map_keys_rule`, `use_falsey_null_rule`, each with tests
+* Script supports arguments: `--stdin` (and alias `-`), `--rules`, and
+  `--stdin-file-url`